admin/requirements.txt
# Admin Console - Standalone Dependencies

# Web Framework
flask>=2.3.0

# LangGraph (for workflow execution)
langgraph>=0.3.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0

# HTTP client (for MCP calls)
httpx>=0.25.0

# Optional: Gemini support
google-generativeai>=0.5.0
